Review fixes: steady "PC Sync", replay-proof handshake, no password in the repo
- Bridge: the "SignalRGB is streaming" timer read a packet stamped in the
  current millisecond as 49 days old, so the pages flashed "WLED effect"
  for a moment every few seconds. Compared signed now.
- Link protocol, light side: a session counts as up only once the bridge's
  first sealed frame proves it is live (keys from the light's fresh nonce),
  and only then does the light follow the announced mode and channel; a
  working session ignores HELLOs. A recorded HELLO played back can no longer
  break the link or flip the light's mode. Self-test: 20/20 on the ESP32.
- Light usermod: fixed a one-byte overflow of the socket buffer; the link
  key moves from cfg.json to NVS (out of WLED's settings pages and backups);
  channel hunting stops only on a real handshake; channel moves in ESP-NOW
  mode retune instead of restarting the radio.
- Bridge: no default Wi-Fi password any more; each bridge makes up its own
  on first start (true randomness, also before the radio is on). Datagrams
  go out without a 1.7 KB stack copy. "Auto channel" no longer forgets the
  saved channel when the rest of the settings are refused.
- PC: a new Wi-Fi password goes to WLED first (bridge-config, and the new
  --new-password), and WLED is kept in step with the bridge over ESP-NOW.
  The control page shows where the phone page lives (this repository's
  GitHub Pages, served through index.html -> pc/phone/).
- README: clearer iPhone/Bluefy steps; the retired password is gone.

Tests: 87/87.

+++ bridge/src/wled.cpp
@@ -375,7 +375,9 @@ static void applyCommand(const Cmd& cmd) {
 // through this bridge, so the bridge can tell by itself.
 static void trackStream() {
   uint32_t at = lastStreamMs;
-  int live = at && millis() - at < STREAM_TIMEOUT_MS ? 1 : 0;
+  // signed: a packet noted in this very millisecond can be 1 ms "ahead" (see wledNoteStream), and unsigned
+  // that read as 49 days ago, so the pages flashed "WLED effect" for a moment every few seconds
+  int live = at && (int32_t)(millis() - at) < (int32_t)STREAM_TIMEOUT_MS ? 1 : 0;
   if (live == state.live) return;  // only this task changes the state, so no lock needed to read it
   State s = snapshot();
   s.live = live;
